Dicegame.m
In this project, we are simulating two person who want to win the dicing game. The winner would be either of them if the total final points of thier throwing is over 50 first. How the rules goes is one of them would keep throwing two dices in one run until he see either of two dices is 1 in this turn and those number are count to 0 this time in this run, then the rest of points in this run could be save into his final point. Finally, if either of them this turn is added up to 50 points, then this person is set to win.  For example, in first run first turn, person a has thrown two dices of 2 and 3. In first run second turn, a has thrown two dices of 1 and 4. Then in this first run, final score of perrson a are set to 5 because 2+3 is 5 and second turn is disregarded. Until we see final point of someone is over 50 then we announce this person has won.